refactor(db): extract inline tests from 10 modules to sibling files
Per .claude/rules/03-unit-testing.md, modules with non-trivial inline
#[cfg(test)] blocks (more than 1-2 short cases) should promote their
tests to a sibling tests.rs file — using db/src/books/tests.rs as the
model.

Extracted the inline #[cfg(test)] mod tests { ... } blocks from 10
db/src/ modules into sibling <module>/tests.rs files. No production
behavior changes; the prod files now end with `#[cfg(test)] mod tests;`
declarations only.

Files extracted:
- db/src/ebook/stat.rs        -> db/src/ebook/stat/tests.rs
- db/src/audiobook/group.rs   -> db/src/audiobook/group/tests.rs
- db/src/auth/password.rs     -> db/src/auth/password/tests.rs
- db/src/auth/users.rs        -> db/src/auth/users/tests.rs
- db/src/auth/session.rs      -> db/src/auth/session/tests.rs
- db/src/ebook/cover.rs       -> db/src/ebook/cover/tests.rs
- db/src/scanner.rs           -> db/src/scanner/tests.rs
- db/src/covers.rs            -> db/src/covers/tests.rs
- db/src/pool.rs              -> db/src/pool/tests.rs
- db/src/author_photos_data.rs -> db/src/author_photos_data/tests.rs

All 592 omnibus-db lib tests pass; cargo fmt + cargo clippy clean.
Existing shared helpers in crate::test_support / crate::ebook::test_support
/ crate::auth::test_support continued to be used — no helper
duplication was introduced.

Closes #413
